% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_outlier.R
\name{adj_outlier}
\alias{adj_outlier}
\title{Corrects data located in outlier.}
\usage{
adj_outlier(Input, Target = 0.01, Winsorizing = TRUE)
}
\arguments{
\item{Input}{Factor Exposure Data}

\item{Target}{reference value, basically 1 percent}

\item{Winsorizing}{TRUE for Winsorizing method, FALSE for Trim(Truncation) method}
}
\description{
It is possible to use the
Windsorizing method to correct the data above the reference value and the
Trim(Truncation) method to delete data above the reference value.
}
\examples{
\dontrun{
Input = sample(-1000:1000, 100, replace=F)
x = adj_outlier(Input, Target = 0.01, Winsorizing = TRUE)
 }
}
