% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_KOR_fs.R
\name{arrange_KOR_fs}
\alias{arrange_KOR_fs}
\title{Arrange all listed firm's financial statement data in US Markets.}
\usage{
arrange_KOR_fs()
}
\value{
arranged financial statement data by account
}
\description{
You should be execute get_US_fs function first to download financial statement data.
It will arrange fs data by account for list type, and save csv file.
}
\examples{
\dontrun{
 KOR_fs = arrange_KOR_fs()
 }
}
