\name{asset_data}
\alias{asset_data}
\docType{data}
\title{Gloal Asset Return Data}
\description{Daily data of ten global asset return data beginning on
1993-01-01 and ending 2018-12-31. The ETF price was used first, and public funds and index prices were used before listing. All returns include dividends.
}
\format{xts type with daily observations}
\details{
The indexes are: US Equity, Europe Equity, Japan Equity, EM Equity, US LongTerm Bond,
US MidTerm Bond, US REITs, Global REITs, Gold, Commodities
}
\examples{
data(asset_data)
}
\keyword{datasets}
