% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_pkg_trend.R
\name{cran_pkg_trend}
\alias{cran_pkg_trend}
\title{Download number of packages listed at CRAN}
\usage{
cran_pkg_trend()
}
\description{
Download number of packages listed at CRAN
}
\examples{
\dontrun{
  cran_pkg_trend()
  }
}
