% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.mom.strength.R
\name{eps.mom.strength}
\alias{eps.mom.strength}
\title{Calculate Strenghness of EPS Momentum}
\usage{
eps.mom.strength(EPS)
}
\arguments{
\item{EPS}{Earnings Per Share}
}
\description{
This function calculates how stable the EPS has grown.
}
\examples{
\dontrun{
EPS = c(3506,2713,2198,2735,5421)
eps.mom.strength(EPS)
 }
}
