% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Score_US.R
\name{F_Score_US}
\alias{F_Score_US}
\title{Calcaulte F-Score of US Firm}
\usage{
F_Score_US(name)
}
\arguments{
\item{name}{the name of arranged Financial statement Data
Result of arrange_US_fs function is US_fs}
}
\value{
All listed firm's F-Score
}
\description{
You should be execute get_US_fs and arrange_US_fs function first
}
\examples{
\dontrun{
 F_US = F_Score_US(US_fs)
 }
}
