% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Score_value_US.R
\name{F_Score_value_US}
\alias{F_Score_value_US}
\title{Return Low Valuation Firm within specific F-Score Firm in U.S}
\usage{
F_Score_value_US(name, score = c(8, 9), n = 30)
}
\arguments{
\item{name}{the name of arranged Financial statement Data}

\item{score}{F-score you want to screen}

\item{n}{Number of stocks you want to invest}
}
\value{
Within Entered F-Score Firm, select top N firm based on valuation ratio.
}
\description{
This function return low valuation firm (Low PBR, Low PER, and Hig Dividend Yield)
in specific F-Score firm (Such as 8 and 9).
You should be execute get_US_fs and arrange_US_fs function first.
}
\examples{
\dontrun{
 invest = F_Score_value_US(US_fs, score = c(8, 9), n = 30)
 }
}
