% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_library.R
\name{french_library}
\alias{french_library}
\title{Download Factor Data of French Library}
\usage{
french_library()
}
\description{
It shows the factors that can be downloaded from the French library.
If you enter a number, corresponding data is downloaded.
}
\examples{
\dontrun{
  data = french_library()
  }
}
