% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Fama_French_Korea.R
\name{get_Fama_French_Korea}
\alias{get_Fama_French_Korea}
\title{Download the Fama-French 5 Factor and Momentum Factor data.}
\usage{
get_Fama_French_Korea(period = "daily")
}
\arguments{
\item{period}{'daily', 'monthly', or 'yearly'}
}
\description{
Initially, you need to authenticate your identity to connect google drive and API.
For more information on generating factor indices, see the following link:
https://docs.google.com/document/d/11A1NH8dy38T4Ay5UDL-ReCCx9VqV3UujeF-imx0ABu4/edit
}
\examples{
\dontrun{
 get_Fama_French_Korea()
 }
}
