% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JAPAN_fs.R
\name{get_JAPAN_fs}
\alias{get_JAPAN_fs}
\title{Download all listed firm's financial statement data in JAPAN Markets.}
\usage{
get_JAPAN_fs()
}
\value{
stock value and financial statement data
}
\description{
This function will Download all listed firm's financial statement data in JAPAN Markets. (NYSE, NASDAQ, AMEX Market)
It will aumomatically save individual stock value, financial statement
and combined data for csv & Rds types
}
\examples{
\dontrun{
 get_JAPAN_fs()
 }
}
