% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JAPAN_price.R
\name{get_JAPAN_price}
\alias{get_JAPAN_price}
\title{Download all listed firm's price in JAPAN Markets.}
\usage{
get_JAPAN_price()
}
\value{
stock prices
}
\description{
This function will Download all listed firm's price in JAPAN Markets. (TOPIX)
It will aumomatically save individual stock prices and
combined prices for csv types
}
\examples{
\dontrun{
 get_JAPAN_price()
 }
}
