% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JAPAN_ticker.R
\name{get_JAPAN_ticker}
\alias{get_JAPAN_ticker}
\title{Download listed firm's ticker in JAPAN(TOPIX) Markets.}
\usage{
get_JAPAN_ticker()
}
\value{
listed company data
}
\description{
This function will Download listed firm's ticker, name, and sector in JAPAN Markets.
}
\examples{
\dontrun{
 get_JAPAN_ticker()
 }
}
