% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_fs2.R
\name{get_KOR_fs2}
\alias{get_KOR_fs2}
\title{This Function is integrated to get_KOR_fs}
\usage{
get_KOR_fs2()
}
\value{
nothing
}
\description{
This Function is integrated to get_KOR_fs
}
\examples{
\dontrun{
  get_KOR_fs2()
  }
}
