% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_sector.R
\name{get_KOR_sector}
\alias{get_KOR_sector}
\title{Download the components of WICS standard 10 sectors.}
\usage{
get_KOR_sector(src = "wics")
}
\arguments{
\item{src}{Download from 'wics'}
}
\description{
Download the components of WICS standard 10 sectors.
}
\examples{
\dontrun{
 sector = get_KOR_sector()
 }
}
