% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_ticker.R
\name{get_KOR_ticker}
\alias{get_KOR_ticker}
\title{Download all listed firm's ticker in KOREA Markets.}
\usage{
get_KOR_ticker(src = "krx")
}
\arguments{
\item{src}{Download from 'krx' or 'naver'}
}
\description{
This function will Download all listed firm's ticker, name in KOR Markets.
}
\examples{
\dontrun{
 ticker = get_KOR_ticker()
 }
}
