% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_value.R
\name{get_KOR_value}
\alias{get_KOR_value}
\title{Download all listed firms valuation data in korea markets}
\usage{
get_KOR_value()
}
\value{
Save valuation data for all listed firms
}
\description{
This function will Download all listed firms valuation data
(EPS, BPS, PER, PEER PER, PBR, Dividend Yield)
}
\examples{
\dontrun{
  get_KOR_value()
  }
}
