% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom_strength.R
\name{get_mom_strength}
\alias{get_mom_strength}
\title{Get Strength of Momentum}
\usage{
get_mom_strength(R, Comparison = 0)
}
\arguments{
\item{R}{Return Data}

\item{Comparison}{The value to be the comparison value'}
}
\description{
It calculate strength of Momentum based on Monthly Return
The comparison value is set to 0 by default and can be changed to the desired value. (Ex: risk free rate)
}
\examples{
\dontrun{
  R = asset_data
  R = R["2000"]
  st_value = get_mom_strength(R, 0)
  }
}
