% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R
\name{get_price}
\alias{get_price}
\title{Download individual firm's price in Korea Markets.}
\usage{
get_price(ticker = "005930", days = 3000)
}
\arguments{
\item{ticker}{symbol for stock}

\item{days}{periods of days (default to 3000)}
}
\description{
This function will Download all listed firm's Adjusted price in Korea Markets.
}
\details{
It will aumomatically save individual stock prices
}
\examples{
\dontrun{
 get_price(ticker = '005930', days = 3000)
 }
}
