% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_bizday.R
\name{get_recent_bizday}
\alias{get_recent_bizday}
\title{Get Recent Business day in Korea Stock Market}
\usage{
get_recent_bizday()
}
\description{
Get Recent Business day in Korea Stock Market
}
\examples{
\dontrun{
 bizday = get_recent_bizday()
 }
}
