% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Table.R
\name{get_Table}
\alias{get_Table}
\title{A function for organizing the names of tickers and items needed for the
 get_kor_fs package.}
\usage{
get_Table(url = url)
}
\arguments{
\item{url}{"http://www.sejongdata.com/query/value.html"}
}
\description{
This function will Download all listed firms financial statement data,
usually data for the last 10 years
}
\details{
It needs "rvest" and "stringi" package.
Scraps from http://www.sejongdata.com
}
\examples{
\dontrun{
  get_kor_fs(folder_name = "Korea_FS", sleep = 2)
  }
}
