% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_fs.R
\name{get_US_fs}
\alias{get_US_fs}
\title{Download all listed firm's financial statement data in US Markets.
This function will Download all listed firm's financial statement data in US Markets. (NYSE, NASDAQ, AMEX Market)
It will aumomatically save individual stock value, financial statement
and combined value for csv types}
\usage{
get_US_fs()
}
\value{
stock value and financial statement data
}
\description{
Download all listed firm's financial statement data in US Markets.
This function will Download all listed firm's financial statement data in US Markets. (NYSE, NASDAQ, AMEX Market)
It will aumomatically save individual stock value, financial statement
and combined value for csv types
}
\examples{
\dontrun{
 get_US_fs()
 }
}
