% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_price.R
\name{get_US_price}
\alias{get_US_price}
\title{Download all listed firm's price in US Markets.}
\usage{
get_US_price(src = "yahoo")
}
\arguments{
\item{src}{"yahoo" or "google"}
}
\value{
stock prices
}
\description{
This function will Download all listed firm's price in US Markets.
(NYSE, NASDAQ, AMEX Market)
}
\details{
If the source is "yahoo", it will download Adjsted price,
and if source is "google", it will download normal price

It will aumomatically save price for csv types
}
\examples{
\dontrun{
 get_US_price(src = "yahoo")
 }
}
