% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_ticker.R
\name{get_US_ticker}
\alias{get_US_ticker}
\title{Download all listed firm's ticker in US Markets.}
\usage{
get_US_ticker()
}
\value{
ticker, name, sector
}
\description{
This function will Download all listed firm's ticker, name, and sector in US Markets.
(NYSE, NASDAQ, AMEX Market)
}
\examples{
\dontrun{
 get_US_ticker()
 }
}
