% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.ratio.R
\name{K.ratio}
\alias{K.ratio}
\title{Caculated Zephyr K-Ratio}
\usage{
K.ratio(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
It calculate Zephyr K-Ratio which is consistency of an equity's return over time
K-Ratio is slope of trend line divided by standard error of trend line
More information can be found at:
http://s3.amazonaws.com/zanran_storage/www.styleadvisor.com/ContentPages/2449998087.pdf
}
\examples{
\dontrun{
  ret = asset_data
  K.ratio(ret[,1])
  }
}
