% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOR_mkt_value.R
\name{KOR_mkt_value}
\alias{KOR_mkt_value}
\title{Download the value index of Korean stock market from Korea Exchange.}
\usage{
KOR_mkt_value(market = "KOSPI", value = "PBR")
}
\arguments{
\item{market}{KRX, KOSPI, KOSDAQ}

\item{value}{PBR, PER, DY}
}
\value{
PBR, PER, and Div Yield indicators of KRX, KOSPI, and KOSDAQ markets.
}
\description{
Download the value index of Korean stock market from Korea Exchange.
}
\examples{
\dontrun{
  KOR_mkt_value(market = "KOSPI", value = "PBR")
  }
}
