% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_list.R
\name{library_list}
\alias{library_list}
\title{List of French Library Data}
\usage{
library_list()
}
\description{
Data From French Library: http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html
}
\examples{
\dontrun{
  list = library_list()
  }
}
