% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative.R
\name{plot_cumulative}
\alias{plot_cumulative}
\title{Plot Cumulative Return}
\usage{
plot_cumulative(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
This function plot Cumulative return
}
\examples{
  R = asset_data
  plot_cumulative(R[,1])
}
