% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drawdown.R
\name{plot_drawdown}
\alias{plot_drawdown}
\title{Plot Drawdown}
\usage{
plot_drawdown(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
This function plot Drawdown
}
\examples{
  R = asset_data
  plot_drawdown(R[,1])
}
