% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Plot Line}
\usage{
plot_line(df, na = "last")
}
\arguments{
\item{df}{Data Frame}

\item{na}{'last' to lastest data, '0 to fill with 0}
}
\description{
This function plot Line Graph
}
