% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_StackedBar.R
\name{plot_StackedBar}
\alias{plot_StackedBar}
\title{Plot Stacked Barchart}
\usage{
plot_StackedBar(df)
}
\arguments{
\item{df}{Time Series data}
}
\description{
This function plot Stacked Barchart
}
\examples{
\dontrun{
  plot_StackedBar(df)
  }
}
