% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_yearly.R
\name{plot_yearly}
\alias{plot_yearly}
\title{Plot yearly Return}
\usage{
plot_yearly(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
This function plot yearly return
}
\examples{
  R = asset_data
  plot_yearly(R)
}
