% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank.IC.R
\name{Rank.IC}
\alias{Rank.IC}
\title{Calculate Cumulative Return Series}
\usage{
Rank.IC(Rank, R)
}
\arguments{
\item{Rank}{Rank Data for Factor}

\item{R}{Return Data}
}
\description{
Calculate Cumulative Return Series
}
