% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reb_point.R
\name{reb_point}
\alias{reb_point}
\title{Return Rebalancing Point by Period}
\usage{
reb_point(R, start, end, period)
}
\arguments{
\item{R}{Return Data}

\item{start}{Starting Date}

\item{end}{Ending Date}

\item{period}{Rebalancing Period by Number}
}
\description{
This function will return rebalacing period, from START to END by REBALANCING PERIOD.
START and END need specific date and Period shold be number.
}
\examples{
\dontrun{
  ret = asset_data
  reb_time = reb_point(ret, "1994-12-31", "2016-12-31", 3)
  }
}
