% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return.CumSeries.R
\name{Return.CumSeries}
\alias{Return.CumSeries}
\title{Calculate Cumulative Return Series}
\usage{
Return.CumSeries(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
Calculate Cumulative Return Series
}
\examples{
  ret = asset_data
  cum_series = Return.CumSeries(ret)
}
