% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return.portfolio.cut.R
\name{Return.portfolio.cut}
\alias{Return.portfolio.cut}
\title{Based on Return.portfolio function, it adds loss cut function.}
\usage{
Return.portfolio.cut(R, weights = NULL, loss_cut = -10,
  rebalance_on = c(NA, "years", "quarters", "months", "weeks", "days"),
  value = 1, verbose = FALSE, ...)
}
\arguments{
\item{R}{An xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{weights}{A time series or single-row matrix/vector containing asset
weights, as decimal percentages, treated as beginning of period weights.  See Details below.}

\item{loss_cut}{Loss Cut point. Default point is -10 to avoid loss cut.}

\item{rebalance_on}{Default "none"; alternatively "daily" "weekly" "monthly" "annual"  to specify calendar-period rebalancing supported by \code{endpoints}.}

\item{value}{The beginning of period total portfolio value. This is used for calculating position value.}

\item{verbose}{If verbose is TRUE, return a list of intermediary calculations.
See Details below.}

\item{\dots}{any other passthru parameters. Not currently used.}
}
\value{
returns a time series of returns weighted by the \code{weights}, applied loss cut function.
}
\description{
If specific assets exceed the Loss Cut limit before rebalancing,
sell all of its assets that exceed the cut point.
Frequency of return should be shorter than frequency of weights.#'
}
\examples{
\dontrun{
data(edhec)
Return.portfolio.cut(edhec["1997",1:5], rebalance_on="quarters", loss_cut = -0.10)
}
}
