% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return.result.R
\name{Return.result}
\alias{Return.result}
\title{Calculate returns & turnover for a portfolio of assets}
\usage{
Return.result(R, weight = NULL, fee = 0)
}
\arguments{
\item{R}{An xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{weight}{A time series or single-row matrix/vector containing asset weights, as decimal percentages, treated as beginning of period weights.}

\item{fee}{Buy/Sell fee, cost, and tax.}
}
\value{
gross is before cost, net is after cost, and turnover is turnover rate
}
\description{
This function using Return.Portfolio function
}
\details{
Using a time series of returns and weights for each asset,
this function calculates the returns of a portfolio with the
Gross Return, Net Return (after fee), and Turnover of Portfolio.
}
\examples{
  ret = asset_data[,c(1,5)]
  portfolio = Return.result(ret)
  port_gross = portfolio$gross
  port_net = portfolio$net
  port_turnover = portfolio$turnover
}
