% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return.stats.R
\name{Return.stats}
\alias{Return.stats}
\title{Statistics of Portfolio Return & Risk}
\usage{
Return.stats(R)
}
\arguments{
\item{R}{Return Data}
}
\description{
It calculate statistics related to portfolio return and risk
This function mostly use PerformanceAnalytics Package
}
\examples{
\dontrun{
  ret = asset_data
  port_stat = Return.stats(ret)
  }
}
