% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskBudget.R
\name{RiskBudget}
\alias{RiskBudget}
\title{Risk Budget Portfolio}
\usage{
RiskBudget(covmat, target)
}
\arguments{
\item{covmat}{Covariance matrix}

\item{target}{Target of Risk Budget. The sum of budget should be 1}
}
\description{
This function solves for risk budge portfolio weights
}
\examples{
\dontrun{
  ret = asset_data
  R = ret[,c(1,5,9)]
  covmat = cov(R)
  target = c(0.4, 0.3, 0.3)
  weight = RiskBudget(covmat, target)
  }
}
