% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_cor.R
\name{roll_cor}
\alias{roll_cor}
\title{Calculate & Plot Rolling Correlation for two assets.}
\usage{
roll_cor(Ra, Rb, width)
}
\arguments{
\item{Ra}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{Rb}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{width}{number of periods to apply rolling function window over}
}
\description{
This function will caculate and plot rolling correlation between
return of two assets.
}
\details{
This Package is based on PerformanceAnalytics
}
\examples{
  ret = asset_data[,c(1,5)]
  roll_cor(ret[,1], ret[,2], width = 360)
}
