% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_minvol.R
\name{wt_minvol}
\alias{wt_minvol}
\title{Minimum Volatility Portfolio}
\usage{
wt_minvol(covmat, lb = NULL, ub = NULL, cut_w = 0.001)
}
\arguments{
\item{covmat}{Covariance matrix}

\item{lb}{Lower weight boundary. Default is Null}

\item{ub}{Upper weight boundary. Default is Null}

\item{cut_w}{If the calculated weight is lower than the corresponding weight,
the weight is made zero. The default value is 1bp.}
}
\description{
This function solves for minimum volatility portfolio weights.
The lower limit and upper limit can be set differently for each asset.
}
\examples{
\dontrun{
  ret = asset_data
  covmat = cov(ret)
  weight = wt_minvol(covmat, lb = rep(0, ncol(covmat)), ub = rep(1, ncol(covmat)))
  }
}
