% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_RiskBudget.R
\name{wt_RiskBudget}
\alias{wt_RiskBudget}
\title{Risk Budget Portfolio}
\usage{
wt_RiskBudget(covmat, target = NULL, optctrl = ctrl(), ...)
}
\arguments{
\item{covmat}{Covariance matrix}

\item{target}{Target of Risk Budget. The sum of budget should be 1}

\item{optctrl}{Object of class Rcpp_CTRL.}

\item{...}{Ellipsis argument is passed down to nlminb().}
}
\description{
This function solves for risk budge portfolio weights
}
\examples{
\dontrun{
  ret = asset_data
  R = ret[,c(1,5,9)]
  covmat = cov(R)
  target = c(0.4, 0.3, 0.3)
  weight = RiskBudget(covmat, target)
  }
}
