% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_fs.R
\name{get_KOR_fs}
\alias{get_KOR_fs}
\title{Download all listed firms financial statement data in korea markets}
\usage{
get_KOR_fs()
}
\value{
Save financial statement data for all listed firms
}
\description{
This function will Download all listed firms financial statement data,
usually data for the last 20 years
}
\examples{
\dontrun{
  get_KOR_fs()
  }
}
