% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kor_fs.R
\name{get_kor_fs}
\alias{get_kor_fs}
\title{Download all listed firms financial statement data in korea markets}
\usage{
get_kor_fs(folder_name = "KoreaFS", sleep_time = 1)
}
\arguments{
\item{folder_name}{Where you want the downloaded information to be stored}

\item{sleep_time}{Sleep time when downloading data}
}
\value{
Save financial statement data for all listed firms
}
\description{
This function will Download all listed firms financial statement data,
usually data for the last 10 years
}
\details{
It needs "rvest" and "stringi" package.
Scraps from http://www.sejongdata.com
}
\examples{
\dontrun{
  get_kor_fs(folder_name = "KoreaFS", sleep_time = 1)
  }
}
