% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_fs.R
\name{get_KOR_fs}
\alias{get_KOR_fs}
\title{Download all listed firms financial statement data in korea markets}
\usage{
get_KOR_fs(src = "fn")
}
\arguments{
\item{src}{Download from fn or yahoo}
}
\value{
Save financial statement & and valuation data for all listed firms
}
\description{
This function will Download all listed firm's financial statement data and
valuation data
}
\details{
Downloading data from Yahoo Finance or Company Guide
}
\examples{
\dontrun{
  get_KOR_fs()
  }
}
