% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KOR_fs2.R
\name{get_KOR_fs2}
\alias{get_KOR_fs2}
\title{Download all listed firms financial statement data in korea markets}
\usage{
get_KOR_fs2()
}
\value{
Save financial statement data for all listed firms
}
\description{
This function will Download all listed valuaation and financial statement data,
usually data for the last 4 years
}
\details{
get_KOR_fs is downloading data from SEJONG DATA, and
get_KOR_fs2 is downloading data from FNGUIDE
}
\examples{
\dontrun{
  get_KOR_fs2()
  }
}
