% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_price.R
\name{get_US_price}
\alias{get_US_price}
\title{Download all listed firm's price in US Markets.}
\usage{
get_US_price()
}
\value{
stock prices
}
\description{
This function will Download all listed firm's price in US Markets.
(NYSE, NASDAQ, AMEX Market)
}
\details{
It will aumomatically save individual stock prices and
combined prices for csv types
}
\examples{
\dontrun{
 US_price = get_US_price()
 }
}
