% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative.R
\name{plot_cumulative}
\alias{plot_cumulative}
\title{Plot Cumulative Return}
\usage{
plot_cumulative(R, ylog = FALSE)
}
\arguments{
\item{R}{Return Data}

\item{ylog}{TRUE/FALSE set the y-axis to logarithmic scale}
}
\description{
This function plot Cumulative return
}
\examples{
 \dontrun{
  R = asset_data
  plot_cumulative(R, ylog = TRUE)
  }
}
